move Terminal to presenters package
